misc: stm401: Fix conversion macros.
The STM32_TO_HOST macro was casting to the wrong data type causing
values to be truncated. Both conversion macros now use the correct size
data type when casting the result value.

Reviewed-on: http://gerrit.mot.com/737093
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>

+++ include/linux/stm401.h
@@ -542,8 +542,8 @@ struct stm_response {
 
 /* The following macros are intended to be called with the stm IRQ handlers */
 /* only and refer to local variables in those functions. */
-#define STM16_TO_HOST(x, buf) ((short) be16_to_cpu(*((u16 *) (buf+(x)))))
-#define STM32_TO_HOST(x, buf) ((short) be32_to_cpu(*((u32 *) (buf+(x)))))
+#define STM16_TO_HOST(x, buf) ((int16_t) be16_to_cpu(*((u16 *) (buf+(x)))))
+#define STM32_TO_HOST(x, buf) ((int32_t) be32_to_cpu(*((u32 *) (buf+(x)))))
 
 struct stm401_quickpeek_message {
 	u8 message;
